Fixed some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/900



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@7608 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/tools/genprops/store.c
@@ -779,7 +779,7 @@ generateData(const char *dataDir) {
 
     trieSize=utrie_serialize(pTrie, trieBlock, sizeof(trieBlock), getFoldedPropsValue, TRUE, &errorCode);
     if(U_FAILURE(errorCode)) {
-        fprintf(stderr, "error: utrie_serialize failed: %s (length %ld)\n", u_errorName(errorCode), trieSize);
+        fprintf(stderr, "error: utrie_serialize failed: %s (length %ld)\n", u_errorName(errorCode), (long)trieSize);
         exit(errorCode);
     }
